WIP splitgraph.commands almost completely removed
