--- splitgraph/commands/__init__.py
@@ -1,9 +1,0 @@
-"""
-Splitgraph public command API
-"""
-
-from splitgraph.commands._drawing import render_tree
-from splitgraph.commands.misc import get_log, cleanup_objects, init, rm
-from splitgraph.commands.mounting import mount
-from splitgraph.commands.publish import publish
-from splitgraph.commands.push_pull import push, pull, clone
